+++ controller/service.js
@@ -1,9 +1,15 @@
 import serviceService from '../services/serviceService'
+import houseService from '../services/houseService'
+import roomService from '../services/roomService'
 
 export const createService = async (req,res,next) => {
     try {
-        const {name, cost, unit,description } = req.body
-        const createdService = await serviceService.createService({name, cost, unit,description})
+        const {name, cost, unit,description,house_id } = req.body
+        const house = await houseService.getHouse(house_id)
+        const isBelong = house.owner_id == req.user.userId
+        if (!isBelong) 
+            return res.status(404).json({msg: "Fail", status: false})
+        const createdService = await serviceService.createService({name, cost, unit,description,house_id})
         res.status(200).json({msg:"Created success", data: createdService, status: true})
     } catch (error) {
         console.log(error)
@@ -14,6 +20,10 @@ export const createService = async (req,res,next) => {
 export const allServices= async(req,res,next) => {
     try {
         const id = req.params.id
+        const house = await houseService.getHouse(id)
+        const isBelong = house.owner_id == req.user.userId
+        if (!isBelong) 
+            return res.status(404).json({msg: "Fail", status: false})
         const services = await serviceService.allServices(id)
         res.status(200).json({msg:"All", data: services, status: true})
     } catch (error) {
@@ -27,6 +37,10 @@ export const getService = async (req,res,next) =>{
         const id = req.params.id
         const service = await serviceService.getService(id)
         if(!service) return res.status(404).json({msg: "Not Found", status: false})
+        const house = await houseService.getHouse(service.house_id)
+        const isBelong = house.owner_id == req.user.userId
+        if (!isBelong) 
+            return res.status(404).json({msg: "Fail", status: false})      
         res.status(200).json({msg:"Found", data: service, status: true})
     } catch (error) {
         console.log(error)
@@ -38,6 +52,12 @@ export const updateService = async(req,res,next) => {
     try {
         const {name, cost, unit,description }= req.body
         const id = req.params.id
+        const service = await serviceService.getService(id)
+        if(!service) return res.status(404).json({msg: "Not Found", status: false})
+        const house = await houseService.getHouse(service.house_id)
+        const isBelong = house.owner_id == req.user.userId
+        if (!isBelong) 
+            return res.status(404).json({msg: "Fail", status: false})      
         const updatedService = await serviceService.updateService({name, cost, unit,description,id})
         if(!updatedService) return res.status(404).json({msg: "Not Found", status: false})
         res.status(200).json({msg:"Updated", data: updatedService, status: true})
@@ -50,6 +70,13 @@ export const updateService = async(req,res,next) => {
 export const deleteService = async(req,res,next) => {
     try {
         const id = req.params.id
+        const service = await serviceService.getService(id)
+        if(!service) 
+            return res.status(404).json({msg: "Not Found", status: false})
+        const house = await houseService.getHouse(service.house_id)
+        const isBelong = house.owner_id == req.user.userId
+        if (!isBelong) 
+            return res.status(404).json({msg: "Fail", status: false})      
         await serviceService.deleteService(id)
         res.status(200).json({msg:"Deleted", status: true})
     } catch (error) {
@@ -61,6 +88,11 @@ export const deleteService = async(req,res,next) => {
 export const servicesInRoom = async (req,res,next) => {
     try {
         const id = req.params.id
+        const room = await roomService.getRoom(id)
+        const house = await houseService.getHouse(room.house_id)
+        const isBelong = house.owner_id == req.user.userId
+        if (!isBelong) 
+            return res.status(404).json({msg: "Fail", status: false})     
         const services = await serviceService.servicesInRoom(id)
         res.status(200).json({msg:"All", data:services, status: true})
     } catch (error) {
@@ -72,6 +104,11 @@ export const servicesInRoom = async (req,res,next) => {
 export const addServiceToRoom = async (req,res,next) => {
     try {
         const room_id = req.params.id
+        const room = await roomService.getRoom(room_id)
+        const house = await houseService.getHouse(room.house_id)
+        const isBelong = house.owner_id == req.user.userId
+        if (!isBelong) 
+            return res.status(404).json({msg: "Fail", status: false})      
         const {id_services} = req.body
         const existedServices = await serviceService.servicesInRoom(room_id)
         for(let x of existedServices) {
@@ -90,7 +127,12 @@ export const addServiceToRoom = async (req,res,next) => {
 export const removeServiceInRoom = async (req, res, next) => {
     try {
         const room_id = req.params.id
-        const {service_id} = req.body
+        const room = await roomService.getRoom(room_id)
+        const house = await houseService.getHouse(room.house_id)
+        const isBelong = house.owner_id == req.user.userId
+        if (!isBelong) 
+            return res.status(404).json({msg: "Fail", status: false})     
+         const {service_id} = req.body
         await serviceService.removeServiceInRoom({room_id, service_id})
         res.status(201).json({msg:"Removed", status: true})
     } catch (error) {
